Add more options for xmins and nextxids
This adds two more options for additional metadata into the slot:

- `include-xmins` This data includes the `catalog_xmin` and `xmin` fields
from the replication slot. The reasons for wanting this generally occurs when
using logical replication with a secondary. Since replication slots don't
get replicated to secondaries, in the event of the failover you need
some mechanism for seeing what records you may have missed. The
xmin/catalog_xmin is useful as you can use it for a lower bound and any
records with a higher xmin could have been missed
- `include-next-xids` This includes sending both the `epoch` and
`nextxid` 32-bit ints. The epoch is a number that increments when xid
rollover happens. This allows you to reconstruct the same txid you get
from the `txid_current` function which accounts for rollover. This can
be really useful to see at a point's time how far behind the slot you
currently are.