Add an optional new output format
This adapts the work done for `write-in-chunks` to add a new output
format, called `msg-per-record` this produces an entire newline
delimited json record for every individual change.

This is useful for large transactions while also making the output
easier to parse. As it currently exists with `write-in-chunks`, the
slot consumer needs to either buffer the entire record or implement more
complex stream parsing to handle the changes.

The new format is outlined in the README, but essentially we add two new
`kinds` of records - `begin` and `commit` and then output the elements
of the `changes` arrary as individual documents.