Replace appendStringInfo with appendStringInfoStrings
Removing pretty-print branching replaced several appendStringInfoString
uses with appendStringInfo, which dramatically cleaned up the output
code at the cost of decoding about 22% slower. We can recover most of
this performance and keep the tidiness by adding a helper function to
run appendStringInfoString for a variable number of strings. We lose
some niceties of the sprintf-style template, but the performance seems
worth it for high-throughput databases.

"22% slower" was measured by populating a pgbench database with scale
factor 5, running 50K transactions, DELETE-ing all the data, then timing
how long it takes to receive all the changes with `pg_recvlogical`. On a
c5.4xlarge EC2 instance backed by a gp2 EBS volume, this takes about 9.2
seconds on this branch compared to about 11.2s on master as of
aa4077a6e05ae3ba2995172623b9928a190f1690.