New wal2json format
I'm proposing a new format to make some wal2json users happier. This
proposal was motivated by issues #70 #71 #72 and PR #65 and various off
list requests. It is a RFC and is not the final format.

Features:

* one JSON per tuple;
* each JSON has an "action" (BEGIN, COMMIT, INSERT, UPDATE, DELETE,
TRUNCATE, MESSAGE);
* one (optional) JSON object for BEGIN/COMMIT;
* "columns" is an array of elements;
* each "columns" element is an object that contains name, type and
value;
* "identity" is an array of elements (REPLICA IDENTITY for UPDATE /
DELETE statements);
* each "identity" element is an object that contains name, type and
value.

This new format solves the big transaction issue. Users can control
transactions using include-transaction (that defaults to true) -- JSONs
will be emit at the beginning and end of the transaction.